api/requirements.txt

# Skiff logging
python-json-logger==0.1.10

# Server
fastapi
uvicorn

# MongoDB driver
motor==3.1.1

# used for testing/formatting
pytest
pytest-env
flake8
black==20.8b1
requests

# used for form parsing
python-multipart

# Allows us to use FastAPI's FileResponse for serving pdfs.
aiofiles

# to work with ontologies
owlready2
rdflib

# Password hashing and JWT
passlib[bcrypt]
python-jose[cryptography]

# Data validation
pydantic[email]

pandas
pdfplumber
